refactor(Calculator): 기존 expression 상태값을 활용하여 처리하고 있던 		      로직을 커스텀 훅으로 이동
